requirements.txt
Django==2.1.7
django-discover-runner==1.0
django-role-permissions==2.2.1
Pillow==6.0.0
pytz==2019.1
six==1.12.0
